Partially address #2201
